test/requirements.txt
# Conduktor Helm Charts Test Framework Dependencies

# CLI framework
click>=8.1.0

# Configuration and validation
pydantic>=2.0.0
pyyaml>=6.0.0

# Rich terminal output
rich>=13.0.0

# Git operations for change detection
gitpython>=3.1.0
